multiple fixes

+++ src/server/pdc_e2o/aws/pdc_backend_s3.h
@@ -6,7 +6,7 @@
 #include "pdc_region.h"
 #include "pdc_client_server_common.h"
 
-int PDC_Server_S3_size(char *aws_s3_location);
+uint64_t PDC_Server_S3_size(char *aws_s3_location);
 int PDC_Server_S3_write(char *aws_s3_location, void *buf, uint64_t write_size);
 int PDC_Server_S3_write_region(char *aws_s3_location, void *buf, uint64_t write_size, uint64_t region_offset,
                                int seek_end);
